zshrc/Starship: Add Bash-like window title function
Bash sets the window title based on your current prompt, but both
zsh and Starship lack that, which then results in an empty title
and in turn makes working with multiple tabs counterproductive
in a way. This change right here benefits from the relevant
Starship documentation page[1] to fix exactly that.

I also added the HOSTNAME variable right into zshrc to avoid
calling the "hostname" executable all the time. This could
save some time and resource when navigating directories in a
low-end hardware while there's physically no change in performance
on high-end hardware.

[1]: https://starship.rs/advanced-config/#change-window-title